refactor: migrate to TypeScript
