+++ backend/package.json
@@ -35,6 +35,7 @@
     "coverage": "nyc report --reporter=text-lcov > coverage.lcov"
   },
   "dependencies": {
+    "@types/fast-levenshtein": "^0.0.4",
     "@types/fs-extra": "^11.0.4",
     "cross-env": "^7.0.3",
     "dotenv-flow": "^4.1.0",
